+++ codex-rs/context-fragments/BUILD.bazel
@@ -1,0 +1,6 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "context-fragments",
+    crate_name = "codex_context_fragments",
+)
