+++ codex-rs/core-skills/src/lib.rs
@@ -7,6 +7,7 @@ mod mention_counts;
 pub mod model;
 pub mod remote;
 pub mod render;
+mod skill_instructions;
 pub mod system;
 
 pub(crate) use invocation_utils::build_implicit_skill_path_indexes;
@@ -29,3 +30,4 @@ pub use render::SkillRenderReport;
 pub use render::build_available_skills;
 pub use render::default_skill_metadata_budget;
 pub use render::render_available_skills_body;
+pub use skill_instructions::SkillInstructions;
