+++ codex-rs/core/src/context/mod.rs
@@ -7,8 +7,6 @@ mod available_skills_instructions;
 mod collaboration_mode_instructions;
 mod contextual_user_message;
 mod environment_context;
-mod fragment;
-mod fragments;
 mod guardian_followup_review_reminder;
 mod hook_additional_context;
 mod image_generation_instructions;
@@ -24,7 +22,6 @@ mod plugin_instructions;
 mod realtime_end_instructions;
 mod realtime_start_instructions;
 mod realtime_start_with_instructions;
-mod skill_instructions;
 mod subagent_notification;
 mod turn_aborted;
 mod user_instructions;
@@ -34,15 +31,16 @@ pub(crate) use approved_command_prefix_saved::ApprovedCommandPrefixSaved;
 pub(crate) use apps_instructions::AppsInstructions;
 pub(crate) use available_plugins_instructions::AvailablePluginsInstructions;
 pub(crate) use available_skills_instructions::AvailableSkillsInstructions;
+pub(crate) use codex_context_fragments::AdditionalContextDeveloperFragment;
+pub(crate) use codex_context_fragments::AdditionalContextUserFragment;
+pub use codex_context_fragments::ContextualUserFragment;
+pub(crate) use codex_context_fragments::FragmentRegistration;
+pub(crate) use codex_context_fragments::FragmentRegistrationProxy;
+pub(crate) use codex_core_skills::SkillInstructions;
 pub(crate) use collaboration_mode_instructions::CollaborationModeInstructions;
 pub(crate) use contextual_user_message::is_contextual_user_fragment;
 pub(crate) use contextual_user_message::parse_visible_hook_prompt_message;
 pub(crate) use environment_context::EnvironmentContext;
-pub use fragment::ContextualUserFragment;
-pub(crate) use fragment::FragmentRegistration;
-pub(crate) use fragment::FragmentRegistrationProxy;
-pub(crate) use fragments::AdditionalContextDeveloperFragment;
-pub(crate) use fragments::AdditionalContextUserFragment;
 pub(crate) use guardian_followup_review_reminder::GuardianFollowupReviewReminder;
 pub(crate) use hook_additional_context::HookAdditionalContext;
 pub(crate) use image_generation_instructions::ImageGenerationInstructions;
@@ -60,7 +58,6 @@ pub(crate) use plugin_instructions::PluginInstructions;
 pub(crate) use realtime_end_instructions::RealtimeEndInstructions;
 pub(crate) use realtime_start_instructions::RealtimeStartInstructions;
 pub(crate) use realtime_start_with_instructions::RealtimeStartWithInstructions;
-pub(crate) use skill_instructions::SkillInstructions;
 pub(crate) use subagent_notification::SubagentNotification;
 pub(crate) use turn_aborted::TurnAborted;
 pub(crate) use user_instructions::UserInstructions;
