+++ codex-rs/prompts/Cargo.toml
@@ -14,6 +14,7 @@ workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
+codex-context-fragments = { workspace = true }
 codex-git-utils = { workspace = true }
 codex-execpolicy = { workspace = true }
 codex-protocol = { workspace = true }
